login and logout function

+++ src/components/SurveyResponse.vue
@@ -2,8 +2,12 @@
     <div class="row">
      <div class="col-4" v-for="(answer,index) in answers" v-bind:key="answer.answers">
         <b-button class="ans" v-on:click="send(answer['.key'])"><h3>Response {{ ++index }}</h3></b-button>
-        <b-img class="delete" @click="deleteans(answer['.key'])" src="https://firebasestorage.googleapis.com/v0/b/ptei-1c8c4.appspot.com/o/photo%2Fdelete.png?alt=media&token=3740d8f6-080b-4508-881d-ecba80e14c47" fluid alt="Responsive image" />
+        <b-img class="delete" @click="deleteans(answer['.key'])" v-b-modal.myModal src="https://firebasestorage.googleapis.com/v0/b/ptei-1c8c4.appspot.com/o/photo%2Fdelete.png?alt=media&token=3740d8f6-080b-4508-881d-ecba80e14c47" fluid alt="Responsive image" />
     </div>
+    <!-- <b-modal id="myModal" centered size="sm">
+        <h2>Confirm Delete?</h2>
+    </b-modal> -->
+
     </div>
 </template>
 
@@ -30,11 +34,29 @@ import { firestore } from "../firebase.js";
           this.$router.push({path:'/loader/'+id})
       },
       deleteans(id){
-            firestore.collection('answers').doc(id).delete().then(() =>{
+          this.$swal({
+  title: 'Are you sure?',
+  text: "You won't be able to revert this!",
+  type: 'warning',
+  showCancelButton: true,
+  confirmButtonColor: '#3085d6',
+  cancelButtonColor: '#d33',
+  confirmButtonText: 'Yes, delete it!'
+}).then((result) => {
+  if (result.value) {
+    this.$swal(
+      'Deleted!',
+      'Your file has been deleted.',
+      'success'
+    )
+              firestore.collection('answers').doc(id).delete().then(() =>{
                 this.answers = this.answers.filter(answer => {
                     return answer['.key'] != id
                 })
             })
+  }
+})
+  
         }
   }
 };
